작성자만 수정, 삭제 권한  부여 그리고 관리자 권한 적용

+++ first_project/src/main/java/com/team5/first_project/comment/controller/CommentController.java
@@ -1,5 +1,6 @@
 package com.team5.first_project.comment.controller;
 
+import com.team5.first_project.board.service.BoardService;
 import com.team5.first_project.comment.dto.CommentRequestDto;
 import com.team5.first_project.comment.dto.CommentResponseDto;
 import com.team5.first_project.comment.service.CommentService;
@@ -7,6 +8,7 @@ import com.team5.first_project.member.entity.Member;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.*;
 public class CommentController {
 
     private final CommentService commentService;
+    private final BoardService boardService;
 
     // 댓글 조회
 //    @GetMapping("/comments")
@@ -31,9 +34,15 @@ public class CommentController {
 
     // 댓글 삭제
     @DeleteMapping("/comments/{id}")
-    public ResponseEntity<Void> deleteComment(@PathVariable("id") long id) {
-        commentService.deleteComment(id);
-        return ResponseEntity.noContent().build(); // 204 No Content 응답을 반환
+    public ResponseEntity<Void> deleteComment(@PathVariable("id") long id,
+                                              HttpSession session) {
+        if (commentService.commentAuthorVerification(id, session) ||
+                boardService.administratorVerification(session)){
+            commentService.deleteComment(id);
+            return ResponseEntity.noContent().build(); // 204 No Content 응답을 반환
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     // 댓글 생성
@@ -49,8 +58,14 @@ public class CommentController {
     // 댓글 수정
     @PostMapping("/comments/{commentId}/edit")
     public String updateComment(@PathVariable("commentId") long id,
-                                @Valid @ModelAttribute CommentRequestDto commentRequestDto) {
-        CommentResponseDto commentResponseDto = commentService.updateComment(id, commentRequestDto);
-        return "redirect:/posts/" + commentResponseDto.getPostId();
+                                @Valid @ModelAttribute CommentRequestDto commentRequestDto,
+                                HttpSession session) {
+        if (commentService.commentAuthorVerification(id, session)){
+            CommentResponseDto commentResponseDto = commentService.updateComment(id, commentRequestDto);
+            return "redirect:/posts/" + commentResponseDto.getPostId();
+        } else {
+            CommentResponseDto commentResponseDto = commentService.findComment(id);
+            return "redirect:/posts/" + commentResponseDto.getPostId();
+        }
     }
 }
