+++ first_project/src/main/java/com/team5/first_project/post/controller/PostController.java
@@ -1,5 +1,6 @@
 package com.team5.first_project.post.controller;
 
+import com.team5.first_project.board.service.BoardService;
 import com.team5.first_project.comment.entity.Comment;
 import com.team5.first_project.comment.service.CommentService;
 import com.team5.first_project.member.entity.Member;
@@ -7,11 +8,14 @@ import com.team5.first_project.post.dto.PostRequestDto;
 import com.team5.first_project.post.dto.PostResponseDto;
 import com.team5.first_project.post.entity.Post;
 import com.team5.first_project.post.service.PostService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -27,6 +31,7 @@ public class PostController {
 
     private final PostService postService;
     private final CommentService commentService;
+    private final BoardService boardService;
 
     // 게시글 생성
     @GetMapping("/posts/create")
@@ -66,9 +71,11 @@ public class PostController {
     // 개별 게시글 조회
     @GetMapping("/posts/{postId}")
     public String getPost(@PathVariable("postId") long id,
-                          Model model) {
+                          Model model,
+                          HttpServletRequest request,
+                          HttpServletResponse response) {
+        postService.updateView(id, request, response);
         Post post = postService.findById(id);
-        postService.updateView(id);
         List<Comment> comments = new PostResponseDto(post).getComments();
         List<Comment> orderComments = commentService.orderComment(comments);
 
@@ -98,10 +105,15 @@ public class PostController {
 
     // 게시글 수정
     @GetMapping("/posts/{postId}/edit")
-    public String editPage(@PathVariable("postId") Long id, Model model) {
-        Post post = postService.findById(id);
-        model.addAttribute("post", new PostResponseDto(post));
-        return "post/editPost";
+    public String editPage(@PathVariable("postId") Long id, Model model,
+                           HttpSession session) {
+        if (postService.postAuthorVerification(id, session)){
+            Post post = postService.findById(id);
+            model.addAttribute("post", new PostResponseDto(post));
+            return "post/editPost";
+        } else {
+            return "redirect:/posts/" +id;
+        }
     }
 
     // @AuthenticationPrincipal
@@ -114,9 +126,14 @@ public class PostController {
 
     // 게시글 삭제
     @DeleteMapping("/posts/{postId}")
-    public ResponseEntity<Void> deletePost(@PathVariable("postId") long id) {
-        postService.delete(id);
-        return ResponseEntity.noContent().build(); // 204 No Content 응답을 반환
+    public ResponseEntity<Void> deletePost(@PathVariable("postId") long id,
+                                           HttpSession session) {
+        if (postService.postAuthorVerification(id, session) || boardService.administratorVerification(session)){
+            postService.delete(id);
+            return ResponseEntity.noContent().build(); // 204 No Content 응답을 반환
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
 }
